+++ Subtitles/Patches/AudioSourcePatch.cs
@@ -1,8 +1,10 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using GameNetcodeStuff;
 using HarmonyLib;
 using UnityEngine;
+using static SubtitlesAPI.SubtitlesAPI;
 
 namespace Subtitles.Patches;
 
@@ -65,23 +67,56 @@ public class AudioSourcePatch
 
     private static void AddSubtitle(AudioClip clip)
     {
-        if (clip?.name is null ||
-            !Constants.Translations.TryGetValue(Path.GetFileNameWithoutExtension(clip.name), out string translation))
+        if (clip?.name is null)
         {
-            if (clip is not null && Plugin.Instance.logSoundNames.Value)
-            {
-                Plugin.ManualLogSource.LogInfo($"No translation for {clip.name}.");
-            }
-
             return;
-        };
-
-        if (Plugin.Instance.logSoundNames.Value)
-        {
-            Plugin.ManualLogSource.LogInfo($"Found translation for {clip.name}!");
         }
 
-        Plugin.Instance.subtitles.Add(new Subtitle(translation));
+        string clipName = Path.GetFileNameWithoutExtension(clip.name);
+
+        if (Localization.Translations.TryGetValue(clipName, out string soundTranslation))
+        {
+            if (Plugin.Instance.logSoundNames.Value)
+            {
+                Plugin.ManualLogSource.LogInfo($"Found translation for {clipName}!");
+            }
+
+            Plugin.Instance.subtitles.Add(FormatSoundTranslation(soundTranslation));
+        }
+        else if (Localization.DialogueTranslations.TryGetValue(clipName, out List<(float, string)> translations))
+        {
+            if (Plugin.Instance.logSoundNames.Value)
+            {
+                Plugin.ManualLogSource.LogInfo($"Found dialogue translation for {clipName}!");
+            }
+
+            foreach ((float startTimestamp, string timedTranslation) in translations)
+            {
+                Plugin.Instance.subtitles.Add(ForamtDialogueTranslation(timedTranslation), startTimestamp);
+            }
+        }
+        else
+        {
+            if (Plugin.Instance.logSoundNames.Value)
+            {
+                Plugin.ManualLogSource.LogInfo($"No translation for {clipName}.");
+            }
+        }
+    }
+
+    private static string FormatSoundTranslation(string translation)
+    {
+        if (translation.StartsWith("[") && translation.EndsWith("]"))
+        {
+            return $"<color=yellow>{translation}<color>";
+        }
+
+        return $"<color=yellow>[{translation}]<color>";
+    }
+
+    private static string ForamtDialogueTranslation(string translation)
+    {
+        return $"<color=blue>{translation}<color>";
     }
 
     private static bool IsInWithinAudiableDisable(AudioSource source, float volume)