+++ Subtitles/Subtitle.cs
@@ -1,6 +1,0 @@
-﻿namespace Subtitles;
-
-public class Subtitle(string text)
-{
-    public string Text { get; set; } = text;
-}